10.1.0 (#47)
* testing external access to quick config methods

* finished external

* opacity should be 0.25 - 1.0

* renamed param

* version bump

* just rely on the catch

* managed updates

* fixed web and better catch

* updated country flags

* added on hover passthrough

* updated spacing

* managed update

* updated speed label and default controls background

* added on hover passthrough

* updated default inactive track

* managed updates

* added captions

* added captions hint

* adding captions (not ready, saving progress)

* note to self

* can remove animation controls

* managed updates

* saving progress

* added smart linking

* testing fullscreen

* added window manager

* added fullscreen toggle

* fullscreen one word

* using maybepop

* cleanup

* managed updates

* commented fullscreen stuff

* added

* more links

* added

* comments

* using ez links

* copy/paste whoops

* updated spacing

* updated link padding

* updated tet background conditional

* using listening builders

* default text opacity -> 0.50

* cleanup

* better background colors

* reverted

* small text

* cleanup

* label -> body

* totally random opacity

* added captions key

* cleanup

* (re)added newline

* fixed snackbar bug

* comments

* works now

* added play/replay button in middle for clarity

* better

* fixed bug on modal theme change

* managed updates

* better material padding

* better theme switch handling

* better theme switching

* whoops

* controls shouldn't always be on for mobile

* better mobile ux

* even betterer

* cleanup

* managed updates

* comments

* apptitle -> appname

* removed unused

* updated switches

* upgraded http

* moved updater to efui

* updated updater

* moved updater to efui

* cleanup

* comments

* removed redundant

* cleanup

* managed updates

* updated menu padding

* scrolling controls resets mobile hover timer

* managed updates

* anim duration is an int

* cleanup

* updated router for anim duration

* anim int/double

* saving progress, almost works

* using icons

* noice

* added scroll hints to the horizontal scrolls

* more visible

* version bump

* excluding icon semantics

* updated logs

* sorted add >> add then sort

* updated logs

* updated for efui 10

* missed a spot

* using const

* removed unused

* managed updates

+++ lib/src/consts/config.dart
@@ -136,8 +136,8 @@ const Color lightOutlineVariant = Color(lightOutlineVariantHex);
 
 //* Design settings *//
 
-/// 500.0
-const double defaultAnimationDuration = 500.0;
+/// 500
+const int defaultAnimationDuration = 500;
 
 /// 0x45 / 255
 const double defaultButtonOutlineOpacity = 0.270588235294;
@@ -207,8 +207,8 @@ const double defaultWordSpacing = 1.0;
 
 // Design settings //
 
-/// 0.0
-const double minAnimationDuration = 0.0;
+/// 0
+const int minAnimationDuration = 0;
 
 // Layout settings //
 
@@ -266,8 +266,8 @@ const double minFontHeight = 1.0;
 
 // Design settings //
 
-/// 1000.0
-const double maxAnimationDuration = 1000.0;
+/// 1000
+const int maxAnimationDuration = 1000;
 
 // Layout settings //
 
@@ -478,10 +478,10 @@ const Map<String, Object> baseEmpathConfig = <String, Object>{
   iconSizeKey: defaultIconSize,
 };
 
-/// base with a more compact layout and hidden scrolls
+/// [baseEmpathConfig] with a more compact layout and hidden scrolls
+/// -- ATTENTION --
 /// For open source consumers: this is Empathetech LLC's config
-/// You have permission to modify this code
-/// You do not have permission to use this config in your app
+/// You do NOT have permission to use this config in your production app
 const Map<String, Object> empathMobileConfig = <String, Object>{
   ...baseEmpathConfig,
 
@@ -491,10 +491,10 @@ const Map<String, Object> empathMobileConfig = <String, Object>{
   hideScrollKey: true,
 };
 
-/// base with a more open layout and visible scrolls
+/// [baseEmpathConfig] with a more open layout and visible scrolls
+/// -- ATTENTION --
 /// For open source consumers: this is Empathetech LLC's config
-/// You have permission to modify this code
-/// You do not have permission to use this config in your app
+/// You do NOT have permission to use this config in your production app
 const Map<String, Object> empathDesktopConfig = <String, Object>{
   ...baseEmpathConfig,
 
@@ -505,7 +505,9 @@ const Map<String, Object> empathDesktopConfig = <String, Object>{
 };
 
 /// [EzConfig.defaults] set to all recommended max values
-/// For open source consumers: you may use this in testing, but not in production
+/// -- ATTENTION --
+/// Open source do NOT have permission to use this config in production apps
+/// Also, this is intended for testing anyway
 final Map<String, Object> empathMaxConfig = <String, Object>{
   ...baseEmpathConfig,
 
@@ -557,7 +559,9 @@ final Map<String, Object> empathMaxConfig = <String, Object>{
 };
 
 /// [EzConfig.defaults] set to all recommended min values
-/// For open source consumers: you may use this in testing, but not in production
+/// -- ATTENTION --
+/// Open source do NOT have permission to use this config in production apps
+/// Also, this is intended for testing anyway
 final Map<String, Object> empathMinConfig = <String, Object>{
   ...baseEmpathConfig,
